GameReviews/requirements.txt
Django
Pillow
gunicorn
whitenoise
dj-database-url
psycopg[binary]
django-ckeditor
